Billbe-Error
The eclipse server was not working properly(Tomcat 9).  The Dispatcherservlet was not working properly. I will make some changes to the server and eclipse configuration to whether it will work.  it was showing no error in the code so I am guessing it's the issue with the server assuming the servlert-related jar files are in the server location.